Refactor bookworm details card
